requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
groq==1.0.0
httpx==0.28.1
aiosqlite==0.20.0
asyncpg==0.31.0
redis[hiredis]==7.2.1
websockets==14.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.9
email-validator==2.1.0
pytest==8.3.0
pytest-asyncio==0.24.0
